Update Russian translation
Current translation is very outdated and misses many keys. This one is updated to the latest version of lang file. 

NOTE: The key `chunkloaders.gui.loaded_chunks` gets a bit out of the GUI frame, but this seems to be an issue with some other translations too, so I left it as-is